Move config endpoints to config-settings slice
