+++ clients/admin-ui/src/features/privacy-experience/form/helpers.tsx
@@ -15,6 +15,7 @@ import { Layer1ButtonOption } from "./constants";
 interface LocationOrLocationGroup {
   selected?: boolean;
   id: string;
+  name: string;
 }
 
 export const getSelectedRegionIds = (
@@ -23,6 +24,9 @@ export const getSelectedRegionIds = (
   allLocations
     ?.filter((loc) => loc.selected)
     .map((loc) => loc.id as PrivacyNoticeRegion) ?? [];
+
+export const getSelectedRegions = (allLocations?: LocationOrLocationGroup[]) =>
+  allLocations?.filter((loc) => loc.selected) ?? [];
 
 export const defaultTranslations: ExperienceTranslationCreate[] = [
   {
