+++ clients/admin-ui/src/types/api/models/TCFPublisherSettings.ts
@@ -1,0 +1,3 @@
+export type TCFPublisherSettings = {
+  publisher_country_code?: string;
+};
